pyproject.toml
[project]
name = "water-agi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "asyncio>=3.4.3",
    "httpx>=0.28.1",
    "mcp[cli]>=1.9.4",
    "pydantic-ai>=0.3.1",
    "python-dotenv>=1.1.0",
]
